requirements.txt
pytorch==1.5.0
tqdm==4.46.1
krippendorff==0.4.0
pytorch-lightning==1.3.0
transformers==3.0.2
sentence-transformers==0.3.3
numpy==1.19.4